Fix compiler warnings

+++ src/engine/client/rocketDataGrid.h
@@ -52,7 +52,7 @@ public:
 	~RocketDataGrid() { }
 	void GetRow( Rocket::Core::StringList& row, const Rocket::Core::String& table, int row_index, const Rocket::Core::StringList& columns )
 	{
-		for ( int i = 0; i < columns.size(); ++i )
+		for ( size_t i = 0; i < columns.size(); ++i )
 		{
 			row.push_back( Rocket_QuakeToRML ( Info_ValueForKey( data[ table ][ row_index ].CString(), columns[ i ].CString() ) ) );
 		}
